Use blocks in flow control statements. Other minor cleanups.  No functional changes.

+++ reconcile/src/com/googlecode/openreconcile/server/Result.java
@@ -45,7 +45,7 @@ public class Result{
 	 *  
 	 */
 	public void decreaseScore(int factor) {
-		score = (score*0.95)/((double)factor);
+		score = (score*0.95)/(factor);
 		if (score > 100){
 			score = (double)99;
 		}
